requirements.txt
click==7.1.2
Jinja2==2.11.2
PyYAML==5.3.1